requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==1.5.0
gast==0.3.3
grpcio==1.29.0
html5lib==0.9999999
importlib-metadata==1.6.1
Markdown==3.2.2
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
protobuf==3.12.2
six==1.15.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
tensornets==0.4.6
termcolor==1.1.0
Werkzeug==1.0.1
xlib==0.21
zipp==3.1.0
